msc3861: validate access_token uniqueness manually
If MAS is in use, server does not generate access tokens for its client
devices. As a result, the access_token value in device table will always be
empty. The unique constraint is enabled for the access_token it does
not allow storing empty values in the database so we had to drop it.
Since we still have old login logic, the constraint is still required,
so we have to check the uniqueness manually.